fixing !gcc5

+++ scripts/linux/install.sh
@@ -43,26 +43,32 @@ echoDots(){
     done
 }
 
-sudo add-apt-repository -y ppa:dns/gnu
-sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
-sudo apt-get update -q
-sudo apt-get install gdebi
-wget http://ci.openframeworks.cc/gcc5/gcc5debs.tar.bz2
-tar xjf gcc5debs.tar.bz2
-rm gcc5debs.tar.bz2
-sudo gdebi -n cpp-5_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n g++-5_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n gcc-5-base_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n libstdc++-5-pic_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n libstdc++-5-dev_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n gcc-5_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n gcc-5-multilib_5.4.1-2ubuntu1~14.04_amd64.deb
-sudo gdebi -n g++-5-multilib_5.4.1-2ubuntu1~14.04_amd64.deb
-rm *.deb
+if [ ! -z ${OPT+x} ]; then
+    sudo add-apt-repository -y ppa:dns/gnu
+    sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
+    sudo apt-get update -q
+    sudo apt-get install gdebi
+    wget http://ci.openframeworks.cc/gcc5/gcc5debs.tar.bz2
+    tar xjf gcc5debs.tar.bz2
+    rm gcc5debs.tar.bz2
+    sudo gdebi -n cpp-5_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n g++-5_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n gcc-5-base_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n libstdc++-5-pic_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n libstdc++-5-dev_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n gcc-5_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n gcc-5-multilib_5.4.1-2ubuntu1~14.04_amd64.deb
+    sudo gdebi -n g++-5-multilib_5.4.1-2ubuntu1~14.04_amd64.deb
+    rm *.deb
 
-sudo apt-get install -y coreutils realpath libxrandr-dev libxinerama-dev libx11-dev libxcursor-dev libxi-dev
-sudo apt-get remove -y --purge g++-4.8
-sudo apt-get autoremove
-sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-5 100
-sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-5 100
-g++ -v
+    sudo apt-get install -y coreutils realpath libxrandr-dev libxinerama-dev libx11-dev libxcursor-dev libxi-dev
+    sudo apt-get remove -y --purge g++-4.8
+    sudo apt-get autoremove
+    sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-5 100
+    sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-5 100
+    g++ -v
+else
+    sudo add-apt-repository -y ppa:dns/gnu
+    sudo apt-get update -q
+    sudo apt-get install -y coreutils realpath libxrandr-dev libxinerama-dev libx11-dev libxcursor-dev libxi-dev
+
